pyproject.toml
[project]
name = "project-template-api"
version = "1.0.0"
description = "Django backend for project-template"
authors = [
    { name = "Aleksandr Gorelik", email = "alx.gorelik@gmail.com" },
]
requires-python = ">=3.14"
dependencies = [
    "daphne==4.2.1",
    "django-allauth[headless,socialaccount]==65.16.1",
    "django-cors-headers==4.9.0",
    "django-solo==2.5.1",
    "django==6.0.4",
    "djangorestframework==3.17.1",
    "pillow==12.2.0",
    "psycopg==3.3.3",
    "psycopg-binary==3.3.3",
    "requests==2.33.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar==6.3.0",
    "flake8==7.3.0",
    "isort==8.0.1",
]

[tool.uv]
package = false